requirements.txt
opencv-python
numpy
pyglet
Pillow